@@ -2,8 +2,7 @@
 build-backend = "mesonpy"
 requires = [
     "meson-python>=0.10.0",
-    "wheel",
-    "oldest-supported-numpy"
+    "numpy>=1.25",
 ]
 
 [project]
@@ -20,10 +19,10 @@ classifiers = [
     "License :: OSI Approved :: MIT License",
     "Natural Language :: English",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
     "Topic :: Software Development",
     "Topic :: Scientific/Engineering",
     "Operating System :: Microsoft :: Windows",
@@ -31,12 +30,11 @@ classifiers = [
     "Operating System :: Unix",
     "Operating System :: MacOS",
 ]
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 dependencies = ["numpy"]
 
 [project.optional-dependencies]
 tests = [
-    "flake8",
     "pytest",
     "pytest-cov",
 ]
